Refactor install controller
